added exceptions and some other things

+++ src/main/kotlin/com/kirill/kotlinblog/service/impl/UserServiceImpl.kt
@@ -13,8 +13,8 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 @Transactional
-class UserServiceImpl(val userRepo : UserRepository,
-                      val roleRepo:RoleRepository,
+class UserServiceImpl(val userRepo: UserRepository,
+                      val roleRepo: RoleRepository,
                       val passwordEncoder: PasswordEncoder) : UserService {
 
     override fun saveRole(role: Role): Role {
@@ -35,21 +35,25 @@ class UserServiceImpl(val userRepo : UserRepository,
     override fun getUsers(): List<User> = userRepo.findAll()
 
     override fun saveUser(user: User): User {
-        //if (userRepo.findByUsername(user.password).username == user.username) {
-            //throw RuntimeException("Username already taken")
-        //}
+        if (userRepo.findByUsername(user.username).equals(user.username)){
+            throw RuntimeException("User with such username already exists")
+        }
+
         user.password = passwordEncoder.encode(user.password)
         user.roles.add(roleRepo.findByRole(ERole.ROLE_USER))
         return userRepo.save(user)
     }
 
     override fun deleteById(id: Long) :ResponseEntity<*>{
+        if (userRepo.findById(id).isEmpty){
+            throw RuntimeException("No user found with such id: $id")
+        }
         userRepo.deleteById(id)
         return ResponseEntity.ok().body("User with id: "+id+"successfully removed")
     }
 
-    override fun update(user: User) {
-        TODO("Not yet implemented")
+    override fun update(user: User): User {
+        return user
     }
 
     override fun addRoleToUser(username: String, roleName: ERole) {
